trait and serializer cleanup
